Adds Svcmd_FreezeMode_f() to g_svcmds.c to add server command "sv freeze" to activate or deactivate the freeze mode.
Also adds "freeze_remaining" to g_main.c to store the remaining time
when the freezing was applied during a timed match. After deactivating
the freeze mode, if this remaining time exists, players will not be
moved to their spawn points and the time will be used as the time to run
with a new countdown.

This mode will work in conjuntion with the tournament mode.

+++ src/p_hud.c
@@ -1670,8 +1670,8 @@ void G_SetStats (edict_t *ent)
 		// kernel: show the timer when there is a timelimit and there are 60 seconds left
 		if (!level.ctb_time && timelimit->value)
 		{
-			int totalTime = timelimit->value * 60;
-			int timeElapsed = level.time - gameStartTime;
+			float totalTime = timelimit->value * 60.0;
+			float timeElapsed = level.time - gameStartTime;
 			int timeLeft = totalTime - timeElapsed;
 
 			if (timeLeft < 60)
